+++ CHANGELOG.md
@@ -21,6 +21,7 @@ For a full diff see [`2.0.0...main`][2.0.0...main].
 - Renamed `FileCouldNotBeParsed::fromFileAndParseError()` to `FileCouldNotBeParsed::atFilePathWithParseError()` ([#1127]), by [@localheinz]
 - Used `Source` instead of `string` when referencing source code ([#1130]), by [@localheinz]
 - Started throwing `FileCouldNotBeParsed` when attempting to collect `Construct`s from a `Finder` yielding `SplFileInfo`s with unparseable source code ([#1131]), by [@localheinz]
+- Renamed `FileCouldNotBeRead::at()` and `FileDoesNotExist::at()` to `FileCouldNotBeRead::atFilePath()` and `FileDoesNotExist::atFilePath()` ([#1132]), by [@localheinz]
 
 ## [`2.0.0`][2.0.0]
 
@@ -251,6 +252,7 @@ For a full diff see [`0.4.0...0.5.0`][0.4.0...0.5.0].
 [#1128]: https://github.com/ergebnis/classy/pull/1128
 [#1130]: https://github.com/ergebnis/classy/pull/1130
 [#1131]: https://github.com/ergebnis/classy/pull/1131
+[#1132]: https://github.com/ergebnis/classy/pull/1132
 
 [@ergebnis]: https://github.com/ergebnis
 [@localheinz]: https://github.com/localheinz